Switch RGBA color storage to bytes

+++ src/Files/VolumeFile.h
@@ -200,13 +200,13 @@ namespace caret {
         void getVoxelColorsForSliceInMap(const int32_t mapIndex,
                                          const VolumeSliceViewPlaneEnum::Enum slicePlane,
                                          const int64_t sliceIndex,
-                                         float* rgbaOut) const;
+                                         uint8_t* rgbaOut) const;
 
         void getVoxelColorInMap(const int64_t i,
                                 const int64_t j,
                                 const int64_t k,
                                 const int64_t mapIndex,
-                                float rgbaOut[4]) const;
+                                uint8_t rgbaOut[4]) const;
         
         void clearVoxelColoringForMap(const int64_t mapIndex);
